Refactor: Move createMockFunction to testHelpers.js
- I extracted `createMockFunction` from `client/goToPath.test.js`
  and moved it to `client/testHelpers.js` to allow for reuse across
  multiple test files.
- I updated `client/goToPath.test.js` to import and use the shared
  `createMockFunction` from `client/testHelpers.js`.
- All tests pass after this change.
